% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_versions.R
\name{compare_versions}
\alias{compare_versions}
\title{Compare package versions}
\usage{
compare_versions(package)
}
\arguments{
\item{package}{character, name of the package}
}
\value{
data.frame
}
\description{
Compare package versions on CRAN and GitHub.
}
\details{
\code{compare_versions()} parses the package's index.html on CRAN and looks for the current version as well as URL to a GitHub repo. If the latter is found, \code{compare_versions()} tries to read to current package version from the DESCRIPTION file in the repo.
}
\examples{
#library(compareVersions)
compare_versions("LexisPlotR")
compare_versions("ggplot2")
compare_versions("dplyr")
}
\author{
Philipp Ottolinger
}

